Fixed python example_1
