+++ gldcore/link/python/examples/example_1/x.csv
@@ -1,12 +1,3 @@
-# file...... x.csv
-# date...... Thu May 14 07:45:58 2020
-# user...... dchassin
-# host...... (null)
-# target.... test 0
-# trigger... (none)
-# interval.. 300
-# limit..... 0
-# timestamp,x
 2018-01-01 00:00:00 UTC,0.825548
 2018-01-01 00:05:00 UTC,1.472663
 2018-01-01 00:10:00 UTC,0.654999
@@ -295,4 +286,3 @@
 2018-01-01 23:45:00 UTC,0.238763
 2018-01-01 23:50:00 UTC,13.651192
 2018-01-01 23:55:00 UTC,1.680191
-# end of tape
